5.1.17 Перенесет рендеринг доски и карточек из main.js в controllers/board.js
