+++ src/main.js
@@ -1,93 +1,12 @@
 import BoardComponent from "./components/board.js";
 import FilterComponent from "./components/filter.js";
-import LoadMoreButtonComponent from "./components/load-more-button.js";
-import NoTasksComponent from "./components/no-tasks.js";
 import SiteMenuComponent from "./components/site-menu.js";
-import SortComponent from "./components/sort.js";
-import TaskComponent from "./components/task.js";
-import TaskEditComponent from "./components/task-edit.js";
-import TasksComponent from "./components/tasks.js";
 import {generateFilters} from "./mock/filter.js";
 import {generateTasks} from "./mock/task.js";
-import {render, replace, remove, RenderPosition} from "./utils/render.js";
+import {render, RenderPosition} from "./utils/render.js";
+import BoardController from "./controllers/board.js";
 
 const TASK_COUNT = 22;
-const SHOWING_TASKS_COUNT_ON_START = 8;
-const SHOWING_TASKS_COUNT_BY_BUTTON = 8;
-
-const renderTask = (taskListElement, task) => {
-  const replaceTaskToEdit = () => {
-    replace(taskEditComponent, taskComponent);
-  };
-
-  const replaceEditToTask = () => {
-    replace(taskComponent, taskEditComponent);
-  };
-
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      replaceEditToTask();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-
-  const taskComponent = new TaskComponent(task);
-  const taskEditComponent = new TaskEditComponent(task);
-
-  taskComponent.setEditButtonClickHandler(()=> {
-    replaceTaskToEdit();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-
-  taskEditComponent.setSubmitHandler((evt) => {
-    evt.preventDefault();
-    replaceEditToTask();
-    document.removeEventListener(`keydown`, onEscKeyDown);
-  });
-
-  render(taskListElement, taskComponent, RenderPosition.BEFOREEND);
-};
-
-const renderBoard = (boardComponent, tasks) => {
-  const isAllTasksArchived = tasks.every((task) => task.isArchive);
-
-  if (isAllTasksArchived) {
-    render(boardComponent.getElement(), new NoTasksComponent(), RenderPosition.BEFOREEND);
-    return;
-  }
-
-  render(boardComponent.getElement(), new SortComponent(), RenderPosition.BEFOREEND);
-  render(boardComponent.getElement(), new TasksComponent(), RenderPosition.BEFOREEND);
-
-  const taskListElement = boardComponent.getElement().querySelector(`.board__tasks`);
-
-  let showingTaskCount = SHOWING_TASKS_COUNT_ON_START;
-  tasks.slice(0, showingTaskCount)
-    .forEach((task) => {
-      renderTask(taskListElement, task);
-    });
-
-  const loadMoreButtonComponent = new LoadMoreButtonComponent();
-  render(boardComponent.getElement(), loadMoreButtonComponent, RenderPosition.BEFOREEND);
-
-  loadMoreButtonComponent.setClickHandler(() => {
-    const prevTaskCount = showingTaskCount;
-    showingTaskCount = showingTaskCount + SHOWING_TASKS_COUNT_BY_BUTTON;
-
-    tasks.slice(prevTaskCount, showingTaskCount)
-      .forEach((task) => renderTask(taskListElement, task));
-
-    if (showingTaskCount >= tasks.length) {
-      remove(loadMoreButtonComponent);
-    }
-  });
-
-};
-
 
 const siteMainElement = document.querySelector(`.main`);
 const siteHeaderElement = siteMainElement.querySelector(`.main__control`);
@@ -99,5 +18,6 @@ render(siteHeaderElement, new SiteMenuComponent(), RenderPosition.BEFOREEND);
 render(siteMainElement, new FilterComponent(filters), RenderPosition.BEFOREEND);
 
 const boardComponent = new BoardComponent();
+const boardController = new BoardController(boardComponent);
 render(siteMainElement, boardComponent, RenderPosition.BEFOREEND);
-renderBoard(boardComponent, tasks);
+boardController.render(tasks);
